Performed some major restructuring to try to generalize the code beyond many of it's original MNAV-only structural assumptions.  This will be an ongoing process and each new sensor we are able to support should stretch us in a more positive direction.
For now I have added support for getting gps data from gpsd (either
running locally or on a network connected server.)  This data is used instead
of the MNAV built in gps.

There is more cleanup and code refactoring that needs to be done, but things
are working so this is a good place for a checkpoint commit.



Former-commit-id: feac3632b4a5b6b20f75444ab22d91c8243ed36b