Some SAS updates: - modify target pitch angle in proportion to throttle to mimic real world   pitch stability and make the aircraft behave more expectedly. - Tune the amount of pitch angle modification to be similar to non-SAS flying.
Add support in the config file for providing kalman filter tuning parameters
(initializing the R, Rw, and P matrices) so we can quickly swap sensors with
out needing to change hard coded parameters and recompile the code.

Add support for a new GCS command that can set the look-at target ned vector,
which also flips the system to ned-vector looking when a command is received.
This works together with a new utility on the ground station that can compute
the NED sun vector, shadow vector, and surface sun glint (i.e. over water)
reflection vectors in real time and update them as time progresses.




Former-commit-id: d838bf08eb1aa17507158b53dabcf99831071eb7