Add a configurable pitot tube scale factor for global calibration.  This located in the APM2 config file, but should be overridden in the main.xml at per-aircraft/installation level.  This scales the 'indicated' airspeed to account for variation in installation, location, etc.
There is an on-the-fly pitot tube scale factor estimation that is logged in
the air data packet that runs on top of this.  This tries to estimate 'true'
airpseed on top of the indicated.

Typically you could fly several flights over a range of conditions and
estimate an average value for the true airspeed compensation and update
the calibration from that number (or refine it over time.)

The calibraiton and sensor is probably not accurate enough too worry too much
about indicated vs. true, but you could crunch the numbers in more detail
if you wanted to try to dial this in more carefully.


Former-commit-id: ced0062390af6fede72aaab7bfc89145ecada531