Remove 'flyto-coord' task support.  circle-coord is a much better choice with more sensible and more configurable behavior.
/navigation/home/ is renamed to /mission/home/

/navigation/circle-target/ is renamed to mission/circle/ (to allow additional
configuration properties under the tree.)

Add a request_task_circle_to_altitude() function which sets up a standard
circle hold, but with a target altitude/heading exit condition, and sets the
autopilot to seek the exit altitude.  The the circle hold then is flown until
the altitude is achieved and then exited automatically on the correct exit
heading.


Former-commit-id: 96fb6d99fdf26381eef35c4314799eaf4a31cccd