Add support for live/in property tree changing of circle configuration including: base bank angle, max bank angle, heading/error gain, target agl, and target speed.
Add supporting structure for additional live configuration parameters under
a single property tree path.  Including for starters an exit altitude and an
exit heading condition (optional.)

Only override target altitude or speed if requested by the task config.  This
allows flying a circle hold at whatever the current altitude and speed is set
to.

Save/restore the prior altitude/speed if we have overridden it for this task.

Add a method to return the config path (in the property tree) for this
circle task.


Former-commit-id: cdea46daa319d4c7055e0a42d6aa967461a8d3bf