Lots of clean ups ...
/navigation/home/ renamed to /mission/home/

Compute an ideal entry altitude when the approach route is built.

Add a 'flare' flag.

Save fcs mode, target altitude, and target speed when task is activated.
Restore these when task is closed.

Fly the entry altitude until we reach the first waypoint, then seek the
glideslope.

Compute how far off target altitude we are, and if it's too much, insert
a circle hold to descend (climb?) to correct altitude before exiting at
the correct exit heading to enter the approach and resuming the approach.


Former-commit-id: 9818310a50cc7dd375e333ed62347b38a7d8799b