A big round of updates!  Hopefully things still pretty much work...
(a) we can point from any lon/lat/elev to any other lon/lat/elev.  The way
the math worked out, I did this in earth centric coordinates so it will
actually be exactly the right direction if we are pointing a long way away...
earth curvature should be properly factored in... so if we point to something
on the earth's surface 3000 miles away, we'll actually be pointing through the
earth a bit, not level with the horizon ... of course we won't be able to  see
anything, but if the object is climbing, once it clears the horizon, we'll be
pointing directly towards it which is I think what we want if we are
interested in RF and light.

(b) we can set any orientation on the base station and the pointing should be
calculated correctly.   Different pan/tilt mechanisms can be wildly different
(there is tilt/roll, for example that a lot of quad copters use.)  So the math
approach is hard wired for our pan tilt, but could be updated or made
configurable for other mechanisms.