Update IMU packet parser to match packet definition changes on remote sending side.
Do a timestamp dance to measure the difference between apm2 clock and
local (linux) host clock and generate a low pass filtered (stable) estimate
of the difference.  Use the apm2 timestamp + diff estimate as the imu
timestamp.  This produces a far more stable (and far more correct) imu
timestamp and dt to pass to the EKF.