In order to meet a short deadline to support two raven boards simultaneously:
Move raven driver to raven1, add a raven2 driver.  This is not good software
engineering, but the timeline is too short.  Eventually I should rewrite the
entire sensor/actuator interface to be C++ class based so I can instatiate
multiple objects of any class and manage them ... however, that's a month of
work probably and I need this today!