When receiving a route and the route size changes, zero all the waypoint coordinates before continuing.
For a survey route submission, set target altitude to survey altitude if the
requested altitude makes sense.

Support route end extension parameter.  The intension for this is to extend
each route leg to give the aircarft more room to turn and reintercept the
next transact before overflying the target area.